Renamed to _solvers.py
