+++ scipy/linalg/_solvers.py
